+++ src/app/process/compare/compare.component.ts
@@ -202,11 +202,6 @@ export class CompareComponent
       needRedraw = true;
     }
 
-    if (changes['inputProcessName'] && this.inputProcessName) {
-      this.process_name = this.inputProcessName;
-      needRedraw = true;
-    }
-
     if (needRedraw && this.canvas) {
       this.drawHoroscope(this.process_name);
     }