+++ src/app/process/enum/process.ts
@@ -17,6 +17,7 @@ export enum ProcessName {
   NativecomparDaily = 'NativecomparDaily',
   Direction = 'Direction',
   QuadrantProcess = 'QuadrantProcess',
+  SolarArc = 'SolarArc',
 }
 
 export namespace ProcessName {
@@ -37,6 +38,7 @@ export namespace ProcessName {
     [ProcessName.NativecomparDaily]: '本命比每日回归',
     [ProcessName.Direction]: '主向推运',
     [ProcessName.QuadrantProcess]: '象限推运',
+    [ProcessName.SolarArc]: '太阳弧',
   };
 
   const pathMap: { [key in ProcessName]: Path } = {
@@ -56,6 +58,7 @@ export namespace ProcessName {
     [ProcessName.NativecomparDaily]: Path.NativecomparDaily,
     [ProcessName.Direction]: Path.Direction,
     [ProcessName.QuadrantProcess]: Path.QuadrantProcess,
+    [ProcessName.SolarArc]: Path.SolarArc,
   };
 
   export function name(process: ProcessName): string {